requirements.txt
Brlapi==0.5.6
Django==1.5.4
GnuPGInterface==0.3.2
Mako==0.5.0
MarkupSafe==0.15
MySQL-python==1.2.4
PAM==0.4.2
PIL==1.1.7
Pillow==2.3.1
PyYAML==3.11
Twisted-Core==11.1.0
Twisted-Names==11.1.0
Twisted-Web==11.1.0
Unidecode==0.04.14
adium-theme-ubuntu==0.3.2
apt-xapian-index==0.44
apturl==0.5.1ubuntu3
argparse==1.2.1
ccsm==0.9.5.92
chardet==2.0.1
command-not-found==0.2.44
compizconfig-python==0.9.5.94
configglue==1.0
debtagshw==0.1
defer==1.0.2
dirspec==3.0.0
## FIXME: could not find svn URL in dependency_links for this package:
distribute==0.6.24dev-r0
dj-database-url==0.3.0
dj-static==0.0.5
django-facebook==5.2.10
django-guardian==1.2.0
django-jfu==2.0.1
django-multiuploader==0.2.9
django-toolbelt==0.0.1
django-userena==1.3.0
duplicity==0.6.18
easy-thumbnails==1.5
gunicorn==18.0
hsaudiotag==1.1.1
html2text==3.200.3
httplib2==0.7.2
jockey==0.9.7
keyring==0.7.1
language-selector==0.1
launchpadlib==1.9.12
lazr.restfulclient==0.12.0
lazr.uri==1.0.3
louis==2.3.0
mutagen==1.22
nvidia-common==0.0.0
oauth==1.0.1
onboard==0.97.1
oneconf==0.2.8.1
pexpect==2.3
piston-mini-client==0.7.2
protobuf==2.4.1
psycopg2==2.5.2
py-mysql2pgsql==0.1.5
pyOpenSSL==0.12
pycrypto==2.4.1
pycups==1.9.61
pycurl==7.19.0
pyechonest==8.0.1
pyinotify==0.9.2
pyserial==2.5
pysmbc==1.0.13
pystache==0.5.3
python-apt==0.8.3ubuntu7
python-dateutil==1.5
python-debian==0.1.21ubuntu1
python-echoprint==0.1
python-magic==0.4.6
python-virtkey==0.60.0
pyxdg==0.19
reportlab==2.5
requests==2.2.1
sessioninstaller==0.0.0
simplejson==2.3.2
six==1.6.1
software-center-aptd-plugins==0.0.0
sorl-thumbnail==11.12.1b
static==1.0.2
system-service==0.1.6
termcolor==1.1.0
ubuntuone-couch==0.3.0
ubuntuone-storage-protocol==3.0.2
ufw==0.31.1-1
unattended-upgrades==0.1
unity-lens-video==0.3.5
unity-scope-video-remote==0.3.5
usb-creator==0.2.23
wadllib==1.3.0
wsgiref==0.1.2
xdiagnose==2.5.2
xkit==0.0.0
zope.interface==3.6.1
